requirements.txt
#Local Descriptors
##opencv-python
scikit-image
##numpy
joblib
##matplotlib
##scikit-learn
streamlit-webrtc
##streamlit
##cvzone

#Deep Learning
#torch==2.7.1+cu118
#torchvision==0.22.1+cu118
#torchaudio==2.7.1+cu118
ultralytics==8.3.158
opencv-python==4.11.0.86
numpy==1.26.4
scikit-learn==1.7.1
mediapipe==0.10.21
cvzone==1.6.1
PyYAML==6.0.2
tqdm==4.67.1
matplotlib==3.10.3
pandas==2.3.0
albumentations==1.3.0
streamlit==1.25.0


# pip install torch==2.7.1+cu118 torchvision==0.22.1+cu118 torchaudio==2.7.1+cu118 --index-url https://download.pytorch.org/whl/cu118
# pip install -r requirements.txt
